Move more headers into graphqlservice/internal
